requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
alembic==1.7.6
anyio==3.6.1
asgiref==3.5.1
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
Authlib==0.15.5
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
decopatch==1.4.10
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.2
Flask-Admin==1.6.0
Flask-Executor==0.10.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
makefun==1.13.1
Mako==1.1.6
MarkupSafe==2.0.1
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.9.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytest==7.1.2
pytest-flask==1.2.0
python-dotenv==0.19.2
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.8
Werkzeug==2.0.3
WTForms==3.0.1
xmlrunner==1.7.7
